DataChain-AI/ai-engine/requirements.txt
flask==2.3.2
flask-cors==4.0.0
gunicorn==20.1.0
numpy==1.25.0
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
transformers==4.30.2
openai==0.27.8
langchain==0.0.225
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.9
pytest==7.4.0
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.4.1
redis==4.6.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.17
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
statsmodels==0.14.0
scipy==1.11.1
nltk==3.8.1
spacy==3.6.0
fastapi==0.100.0
uvicorn==0.22.0 